+++ src/str-contract.ts
@@ -33,6 +33,13 @@ export interface StrContract {
   isString (input?: any): input is string
 
   /**
+   * Determine whether the given `input` is an alpha-numeric string.
+   *
+   * @param input
+   */
+  isAlphaNumeric (input?: any): boolean
+
+  /**
    * Determine whether the given `input` is a symbol.
    *
    * @param input - the `input` value to check whether it’s a symbol